Add new reward scaling mechanism
The complicated formulas for `lambda_reward` and
`reward_to_value_factor` are taking into account the fact that not every
time step is an episode start.

`lambda_reward` replaces the reward stddev in normalization, targeting a
true returns stddev of 1 for randomly sampled time steps.

`reward_to_value_factor` is used to convert the mean of rewards to a
mean of returns, which is necessary to unnormalize nn outputs